+++ src/cognitect/aws/endpoint.clj
@@ -30,28 +30,28 @@
 
 (defn service-resolve
   "Resolve the endpoint for the given service."
-  [partition service-name service endpoint-key]
-  (let [endpoint      (get-in service [:endpoints endpoint-key])
-        endpoint-name (name endpoint-key)
-        result        (merge (:defaults partition)
-                             (:defaults service)
-                             endpoint
-                             {:partition     (:partition partition)
-                              :endpoint-name endpoint-name
-                              :dnsSuffix     (:dnsSuffix partition)})]
+  [partition service-name service region-key]
+  (let [endpoint (get-in service [:endpoints region-key])
+        region   (name region-key)
+        result   (merge (:defaults partition)
+                        (:defaults service)
+                        endpoint
+                        {:partition (:partition partition)
+                         :region    region
+                         :dnsSuffix (:dnsSuffix partition)})]
     (util/map-vals #(render-template % {"service"   service-name
-                                        "region"    endpoint-name
+                                        "region"    region
                                         "dnsSuffix" (:dnsSuffix partition)})
                    result
                    [:hostname :sslCommonName])))
 
 (defn partition-resolve
   [{:keys [services] :as partition} service-key region-key]
-  (let [{:keys [partitionEndpoint isRegionalized] :as service} (get services service-key)
-        endpoint-key (if (and partitionEndpoint (not isRegionalized))
-                       (keyword partitionEndpoint)
-                       region-key)]
-    (when (contains? (-> partition :regions keys set) (keyword region-key))
+  (when (contains? (-> partition :regions keys set) region-key)
+    (let [{:keys [partitionEndpoint isRegionalized] :as service} (get services service-key)
+          endpoint-key (if (and partitionEndpoint (not isRegionalized))
+                         (keyword partitionEndpoint)
+                         region-key)]
       (service-resolve partition (name service-key) service endpoint-key))))
 
 (defn resolve*
@@ -63,14 +63,14 @@
   Return a map with the following keys:
 
   :partition            The name of the partition.
-  :endpoint-name        The name of the endpoint.
+  :region               The region of the endpoint.
   :hostname             The hostname to use.
   :ssl-common-name      The ssl-common-name to use (optional).
   :credential-scope     The Signature v4 credential scope (optional).
   :signature-versions   A list of possible signature versions (optional).
   :protocols            A list of supported protocols."
-  [service region]
-  (some #(partition-resolve % service region)
+  [service-key region]
+  (some #(partition-resolve % service-key region)
         (:partitions (resolver))))
 
 (def resolve (memoize resolve*))