+++ test/src/cognitect/aws/signers_test.clj
@@ -103,7 +103,7 @@
         (let [service        {:metadata {:signatureVersion "v4"
                                          :endpointPrefix   "service"
                                          :uid              "service-2018-12-28"}}
-              signed-request (client/sign-http-request service :us-east-1 credentials request)]
+              signed-request (client/sign-http-request service {:region "us-east-1"} credentials request)]
           (is (= (get-in signed-request [:headers "authorization"])
                  authorization))))
       (testing "using signingName"
@@ -111,19 +111,9 @@
                                          :endpointPrefix   "incorrect"
                                          :signingName      "service"
                                          :uid              "service-2018-12-28"}}
-              signed-request (client/sign-http-request service :us-east-1 credentials request)]
+              signed-request (client/sign-http-request service {:region "us-east-1"} credentials request)]
           (is (= (get-in signed-request [:headers "authorization"])
-                 authorization))))
-      (testing "global endpoint is always signed with us-east-1"
-        (let [service {:metadata {:signatureVersion "v4"
-                                  :endpointPrefix   "service"
-                                  :uid              "service-2018-12-28"}}]
-          (let [signed-request (client/sign-http-request service :us-west-1 credentials request)]
-            (is (not= authorization (get-in signed-request [:headers "authorization"]))))
-          (let [signed-request (client/sign-http-request
-                                (assoc-in service [:metadata :globalEndpoint] "the-world")
-                                :us-west-1 credentials request)]
-            (is (= authorization (get-in signed-request [:headers "authorization"])))))))))
+                 authorization)))))))
 
 (deftest test-canonical-query-string
   (testing "ordering"