+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## DEV
 
 * Reduce noise from reflection warnings in java 9+ [#106](https://github.com/cognitect-labs/aws-api/issues/106)
+* Get signing region from endpoint config [#105](https://github.com/cognitect-labs/aws-api/issues/105)
 
 ## 0.8.391 / 2019-10-25